+++ backends/builder.go
@@ -108,18 +108,18 @@ type Builder interface {
 
 // ConvolveAxesConfig defines the interpretation of the input/kernel/output tensor axes.
 // There must be the same number of spatial dimensions (axes) for each of the 3 tensors.
-// Input and output has batch and channel axes. Kernel has inputChannel and outputChannel axes.
+// Input and output have batch and channel axes. Kernel has inputChannel and outputChannel axes.
 //
 // See Builder.ConvGeneralDilated
 type ConvolveAxesConfig struct {
-	InputBatch, InputChannel int
-	InputSpatial             []int
+	InputBatch, InputChannels int
+	InputSpatial              []int
 
-	FilterInputChannel, FilterOutputChannel int
-	FilterSpatial                           []int
+	KernelInputChannels, KernelOutputChannels int
+	KernelSpatial                             []int
 
-	OutputBatch, OutputChannel int
-	OutputSpatial              []int
+	OutputBatch, OutputChannels int
+	OutputSpatial               []int
 }
 
 // Clone returns a deep copy of the structure.
@@ -127,7 +127,7 @@ func (c ConvolveAxesConfig) Clone() ConvolveAxesConfig {
 	var c2 ConvolveAxesConfig
 	c2 = c
 	c2.InputSpatial = slices.Clone(c.InputSpatial)
-	c2.FilterSpatial = slices.Clone(c.FilterSpatial)
+	c2.KernelSpatial = slices.Clone(c.KernelSpatial)
 	c2.OutputSpatial = slices.Clone(c.OutputSpatial)
 	return c2
 }